perf: add benchmark for check function
