+++ Dockerfile
@@ -5,9 +5,10 @@ RUN rustup target add x86_64-unknown-linux-musl
 RUN USER=root cargo new dotenv-linter
 WORKDIR /usr/src/dotenv-linter
 COPY Cargo.toml ./
-RUN cargo build --release
-
 COPY src ./src
+COPY benches ./benches
+
+RUN cargo build --release
 RUN cargo install --target x86_64-unknown-linux-musl --path .
 
 FROM scratch
