+++ benches/fixtures/simple.env
@@ -1,0 +1,10 @@
+ SPACED=
+
+KEY = VALUE
+
+SECRET="my value"
+
+SECRET=Already defined
+
+kebab-case-name=1
+snake_case_name=2
